updated stats from integer to float

+++ db/migrate/20191127222227_create_stats.rb
@@ -2,26 +2,26 @@ class CreateStats < ActiveRecord::Migration[6.0]
   def change
     create_table :stats do |t|
       t.belongs_to :champion, null: false, foreign_key: true
-      t.integer :hp
-      t.integer :hpperlevel
-      t.integer :mp
-      t.integer :mpperlevel
-      t.integer :movespeed
-      t.integer :armor
-      t.integer :armorperlevel
-      t.integer :spellblock
-      t.integer :spellblockperlevel
-      t.integer :attackrange
-      t.integer :hpregen
-      t.integer :hpregenperlevel
-      t.integer :mpregen
-      t.integer :mpregenperlevel
-      t.integer :crit
-      t.integer :critperlevel
-      t.integer :attackdamage
-      t.integer :attackdamageperlevel
-      t.integer :attackspeedperlevel
-      t.integer :attackspeed
+      t.float :hp
+      t.float :hpperlevel
+      t.float :mp
+      t.float :mpperlevel
+      t.float :movespeed
+      t.float :armor
+      t.float :armorperlevel
+      t.float :spellblock
+      t.float :spellblockperlevel
+      t.float :attackrange
+      t.float :hpregen
+      t.float :hpregenperlevel
+      t.float :mpregen
+      t.float :mpregenperlevel
+      t.float :crit
+      t.float :critperlevel
+      t.float :attackdamage
+      t.float :attackdamageperlevel
+      t.float :attackspeedperlevel
+      t.float :attackspeed
 
       t.timestamps
     end
